sentiment-assessment-server/requirements.txt
click==7.1.2
joblib==1.0.0
nltk==3.5
python-dotenv==0.15.0
regex==2020.11.13
textblob==0.15.3
tqdm==4.54.1
websockets==8.1
